final-project/tests/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt tests/requirements.in
#
allure-pytest==2.9.45
    # via -r tests/requirements.in
allure-python-commons==2.9.45
    # via allure-pytest
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.2.0
    # via
    #   allure-python-commons
    #   outcome
    #   pytest
    #   trio
certifi==2021.10.8
    # via
    #   requests
    #   urllib3
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.9
    # via requests
cryptography==36.0.1
    # via
    #   pyopenssl
    #   urllib3
faker==9.8.0
    # via -r tests/requirements.in
h11==0.12.0
    # via wsproto
idna==3.3
    # via
    #   requests
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via pytest
outcome==1.1.0
    # via trio
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via
    #   allure-python-commons
    #   pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pymysql==1.0.2
    # via -r tests/requirements.in
pyopenssl==21.0.0
    # via urllib3
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via
    #   -r tests/requirements.in
    #   allure-pytest
python-dateutil==2.8.2
    # via faker
requests==2.26.0
    # via -r tests/requirements.in
selenium==4.0.0
    # via -r tests/requirements.in
six==1.16.0
    # via
    #   allure-pytest
    #   allure-python-commons
    #   pyopenssl
    #   python-dateutil
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[secure]==1.26.7
    # via
    #   requests
    #   selenium
wsproto==1.0.0
    # via trio-websocket
